improve the data store system
